requirements.txt
compress_fasttext==0.1.5
nltk==3.9.1
numpy==2.2.1
pandas==2.2.3
psycopg2==2.9.10
psycopg2_binary==2.9.10
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.67.1
